cache_single_variation: add cli for most parameters
Refactored this code to make most of the global variables
that tend to change from run to run settable on the command
line with reasonable defaults. Note that stylistically
I just convert these to globals using the old name when
possible to minimize the number of changed lines (for now).

Also fixed a bug where the . and .. directories would
cause two empty batches to be added to each cached output.
Now any file or directory beginning with a dot is ignore.

Note that imwidth is a placeholder and it's best for
now to keep that at 150.