visualize_networks: added cli and logic for different step sizes
Now the number of variations rendered and the range of the
modified parameter can be set from the command line. Backwards
compatible with previous behavior, which could now optionally
be specified as:

  th visualize_networks.lua --num_steps 6 --lower_bound -20 --upper_bound 20

There is also a back-door setting called --custom_bounds which
overrides lower and upper bounds to use an internal hard-coded
table that allows different bounds for different variables.

--imwidth is an untested placeholder, so best for now to keep
that at 150.